+++ apps/api/src/presentation/authentication/authentication.module.ts
@@ -56,6 +56,7 @@ const environmentProvider = {
     AuthenticationTokenService,
     AccessTokenGuard,
     RolesGuard,
+    PermissionsGuard,
     AUTHENTICATION_REPOSITORY,
     EmailVerificationPolicyService,
   ],